Added authentication_needed and get_active_user dependencies. Added currenct user functionallity in some endpoints with POST, DELETE methods. Removed some DELETE, POST endpoints

+++ fast_witter/routers/comments/post_comments.py
@@ -3,12 +3,13 @@ from sqlalchemy.orm import Session
 
 import exc
 import schemas
+from database import models
 
 from database.interfaces.comment_interface import CommentInterface
 from database.interfaces.user_interface import UserInterface
 from database.interfaces.post_interface import PostInterface
 
-from dependencies import get_db, PaginationQueryParams
+from dependencies import get_db, PaginationQueryParams, get_active_user
 
 router = APIRouter(
     prefix='/users/{user_id}/posts/{post_id}/comments',
@@ -48,13 +49,14 @@ def get_post_comments(
     status_code=status.HTTP_201_CREATED
 )
 def create_post_comment(
-        user_id: int, post_id: int,
-        comment: schemas.CommentCreate, db: Session = Depends()
+        user_id: int, post_id: int, comment: schemas.CommentCreate,
+        db: Session = Depends(),
+        active_user: models.User = Depends(get_active_user)
 ):
     raise_exc_if_user_post_not_exist(user_id, post_id, db)
 
     comment = CommentInterface.create_comment(
-        db, user_id, post_id, comment
+        db, active_user.id, post_id, comment
     )
 
     return comment
@@ -91,8 +93,12 @@ def get_post_comment(
 )
 def delete_post_comment(
         user_id: int, post_id: int, comment_id: int,
-        db: Session = Depends(get_db)
+        db: Session = Depends(get_db),
+        active_user: models.User = Depends(get_active_user)
 ):
     comment = get_ind_comment_or_raise_exc(user_id, post_id, comment_id, db)
+
+    if active_user.id != comment.user_id or active_user.id != user_id:
+        raise exc.NotObjectOwner('Comment')
 
     CommentInterface.delete_comment(db, comment)
